requirements.txt
Django>=4.2.11
httpx
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
